@@ -1,6 +1,6 @@
 package org.wit.hillfort.views.hillfortlist
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -14,7 +14,7 @@ interface HillfortListener {
   fun onHillfortClick(hillfort: HillfortModel)
 }
 
-class HillfortAdapter constructor(private var hillforts: List<HillfortModel>, private val listener: HillfortListener) : RecyclerView.Adapter<HillfortAdapter.MainHolder>() {
+class HillfortAdapter constructor(private var hillforts: List<HillfortModel>, private val listener: HillfortListener) : androidx.recyclerview.widget.RecyclerView.Adapter<HillfortAdapter.MainHolder>() {
 
   override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
     return MainHolder(LayoutInflater.from(parent?.context).inflate(R.layout.card_hillfort, parent, false))
@@ -28,7 +28,7 @@ class HillfortAdapter constructor(private var hillforts: List<HillfortModel>, pr
 
   override fun getItemCount(): Int = hillforts.size
 
-  class MainHolder constructor(itemView: View) : RecyclerView.ViewHolder(itemView) {
+  class MainHolder constructor(itemView: View) : androidx.recyclerview.widget.RecyclerView.ViewHolder(itemView) {
 
     fun bind(hillfort: HillfortModel, listener : HillfortListener) {
       itemView.hillfortTitle.text = hillfort.title
